requirements.txt
django-tenants==3.5.0
APScheduler==3.10.1
boto3==1.33.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
django-filter==23.4
django-cors-headers==4.3.1
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
redis==5.0.1
huey==2.5.0
psycopg2==2.9.9
django-choices==2.0.0
phonenumbers==8.13.26
openpyxl==3.1.2
pdfkit==1.0.0
Django==4.2.8
docxtpl==0.16.7
drf-writable-nested==0.7.0
docusign-esign==3.25.0
twilio==8.10.3
Pillow==10.1.0